requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
dashscope>=1.14.0
pydantic>=2.5.0
requests>=2.31.0
